extracting virtualenv files
